# openapi.yaml
...
openapi: 3.1.0
info:
  title: 大模型安全评测系统 API
  description: 大模型安全评测系统后端接口骨架。本服务默认监听本机，不包含登录、用户、角色权限和多租户。
  version: 1.0.0
paths:
...
  /api/v1/risk-taxonomies:
    get:
      tags:
      - 风险分类
      summary: List Risk Taxonomies
      operationId: listRiskTaxonomies
      security:
      - LocalToken: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaginatedData_RiskTaxonomy__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - 风险分类
      summary: Create Risk Taxonomy
      operationId: createRiskTaxonomy
      security:
      - LocalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskTaxonomyCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_RiskTaxonomy_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/risk-taxonomies/{taxonomy_id}/categories:
    get:
      tags:
      - 风险分类
      summary: List Risk Categories
      operationId: listRiskCategories
      security:
      - LocalToken: []
      parameters:
      - name: taxonomy_id
        in: path
        required: true
        schema:
          type: integer
          title: Taxonomy Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_list_RiskCategory__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - 风险分类
      summary: Create Risk Category
      operationId: createRiskCategory
      security:
      - LocalToken: []
      parameters:
      - name: taxonomy_id
        in: path
        required: true
        schema:
          type: integer
          title: Taxonomy Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskCategoryCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_RiskCategory_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/task-attempts/{attempt_id}/risk:
    get:
      tags:
      - 风险量化
      summary: Get Task Attempt Risk
      operationId: getTaskAttemptRisk
      security:
      - LocalToken: []
      parameters:
      - name: attempt_id
        in: path
        required: true
        schema:
          type: integer
          title: Attempt Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_RiskAssessment_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/evaluation-tasks/{task_id}/risk-summary:
    get:
      tags:
      - 风险量化
      summary: Get Risk Summary
      operationId: getRiskSummary
      security:
      - LocalToken: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      - name: risk_level
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Risk Level
      - name: risk_category_code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Risk Category Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_dict_str__Any__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/report-templates:
    get:
      tags:
      - 报告模板
      summary: List Report Templates
      operationId: listReportTemplates
      security:
      - LocalToken: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaginatedData_ReportTemplate__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - 报告模板
      summary: Create Report Template
      operationId: createReportTemplate
      security:
      - LocalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportTemplateCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ReportTemplate_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/report-templates/{template_id}/versions:
    get:
      tags:
      - 报告模板
      summary: List Report Template Versions
      operationId: listReportTemplateVersions
      security:
      - LocalToken: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: integer
          title: Template Id
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaginatedData_ReportTemplateVersion__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - 报告模板
      summary: Create Report Template Version
      operationId: createReportTemplateVersion
      security:
      - LocalToken: []
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: integer
          title: Template Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportTemplateVersionCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ReportTemplateVersion_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/report-template-versions/{version_id}:
    get:
      tags:
      - 报告模板
      summary: Get Report Template Version
      operationId: getReportTemplateVersion
      security:
      - LocalToken: []
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: integer
          title: Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ReportTemplateVersion_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/report-template-versions/{version_id}/publish:
    post:
      tags:
      - 报告模板
      summary: Publish Report Template Version
      operationId: publishReportTemplateVersion
      security:
      - LocalToken: []
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: integer
          title: Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ReportTemplateVersion_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiResponse_PaginatedData_ReportTemplateVersion__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/PaginatedData_ReportTemplateVersion_'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[PaginatedData[ReportTemplateVersion]]
    ApiResponse_PaginatedData_ReportTemplate__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/PaginatedData_ReportTemplate_'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[PaginatedData[ReportTemplate]]
    ApiResponse_PaginatedData_RiskTaxonomy__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/PaginatedData_RiskTaxonomy_'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[PaginatedData[RiskTaxonomy]]
...
    ApiResponse_ReportTemplateVersion_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/ReportTemplateVersion'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[ReportTemplateVersion]
    ApiResponse_ReportTemplate_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/ReportTemplate'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[ReportTemplate]
...
    ApiResponse_RiskAssessment_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/RiskAssessment'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[RiskAssessment]
    ApiResponse_RiskCategory_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/RiskCategory'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[RiskCategory]
...
    ApiResponse_RiskTaxonomy_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/RiskTaxonomy'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[RiskTaxonomy]
...
    ApiResponse_dict_str__Any__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Data
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[dict[str, Any]]
...
    ApiResponse_list_RiskCategory__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/RiskCategory'
            type: array
          - type: 'null'
          title: Data
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[list[RiskCategory]]
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginatedData_ReportTemplateVersion_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ReportTemplateVersion'
          type: array
          title: Items
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
          default: 20
        total:
          type: integer
          minimum: 0.0
          title: Total
          default: 0
        has_next:
          type: boolean
          title: Has Next
          default: false
      type: object
      title: PaginatedData[ReportTemplateVersion]
    PaginatedData_ReportTemplate_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ReportTemplate'
          type: array
          title: Items
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
          default: 20
        total:
          type: integer
          minimum: 0.0
          title: Total
          default: 0
        has_next:
          type: boolean
          title: Has Next
          default: false
      type: object
      title: PaginatedData[ReportTemplate]
    PaginatedData_RiskTaxonomy_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/RiskTaxonomy'
          type: array
          title: Items
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
          default: 20
        total:
          type: integer
          minimum: 0.0
          title: Total
          default: 0
        has_next:
          type: boolean
          title: Has Next
          default: false
      type: object
      title: PaginatedData[RiskTaxonomy]
...
    ReportFormat:
      type: string
      enum:
      - json
      - md
      - csv
      - html
      - pdf
      title: ReportFormat
...
    ReportTemplate:
      properties:
        template_id:
          type: integer
          title: Template Id
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        status:
          type: string
          title: Status
          default: active
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - template_id
      - code
      - name
      title: ReportTemplate
    ReportTemplateCreateRequest:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - code
      - name
      title: ReportTemplateCreateRequest
    ReportTemplateVersion:
      properties:
        template_version_id:
          type: integer
          title: Template Version Id
        template_id:
          type: integer
          title: Template Id
        version:
          type: string
          title: Version
        format:
          type: string
          title: Format
        content:
          type: string
          title: Content
        variables_schema:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Variables Schema
        status:
          type: string
          title: Status
          default: draft
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - template_version_id
      - template_id
      - version
      - format
      - content
      title: ReportTemplateVersion
    ReportTemplateVersionCreateRequest:
      properties:
        version:
          type: string
          title: Version
        format:
          $ref: '#/components/schemas/ReportFormat'
        content:
          type: string
          title: Content
        variables_schema:
          additionalProperties: true
          type: object
          title: Variables Schema
        status:
          type: string
          title: Status
          default: draft
      type: object
      required:
      - version
      - format
      - content
      title: ReportTemplateVersionCreateRequest
...
    RiskAssessment:
      properties:
        attempt_id:
          type: integer
          title: Attempt Id
        overall_score:
          type: number
          maximum: 100.0
          minimum: 0.0
          title: Overall Score
        risk_level:
          $ref: '#/components/schemas/RiskLevel'
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
        uncertainty:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Uncertainty
        judge_trust_score:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Judge Trust Score
        score_version:
          type: string
          title: Score Version
        dimensions:
          items:
            $ref: '#/components/schemas/RiskDimension'
          type: array
          title: Dimensions
        rule_override_applied:
          type: boolean
          title: Rule Override Applied
          default: false
        calculated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Calculated At
      type: object
      required:
      - attempt_id
      - overall_score
      - risk_level
      - confidence
      - uncertainty
      - judge_trust_score
      - score_version
      title: RiskAssessment
    RiskCategory:
      properties:
        risk_category_id:
          type: integer
          title: Risk Category Id
        taxonomy_id:
          type: integer
          title: Taxonomy Id
        parent_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Parent Id
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        definition:
          anyOf:
          - type: string
          - type: 'null'
          title: Definition
        severity_weight:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Severity Weight
          default: 0
        children:
          items:
            $ref: '#/components/schemas/RiskCategory'
          type: array
          title: Children
      type: object
      required:
      - risk_category_id
      - taxonomy_id
      - code
      - name
      title: RiskCategory
    RiskCategoryCreateRequest:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        parent_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Parent Id
        definition:
          anyOf:
          - type: string
          - type: 'null'
          title: Definition
        severity_weight:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Severity Weight
          default: 0
        examples:
          additionalProperties: true
          type: object
          title: Examples
        sort_order:
          type: integer
          title: Sort Order
          default: 0
      type: object
      required:
      - code
      - name
      title: RiskCategoryCreateRequest
...
    RiskLevel:
      type: string
      enum:
      - info
      - low
      - medium
      - high
      - critical
      title: RiskLevel
...
    RiskTaxonomy:
      properties:
        risk_taxonomy_id:
          type: integer
          title: Risk Taxonomy Id
        name:
          type: string
          title: Name
        version:
          type: string
          title: Version
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        is_default:
          type: boolean
          title: Is Default
          default: false
        status:
          type: string
          title: Status
          default: active
      type: object
      required:
      - risk_taxonomy_id
      - name
      - version
      title: RiskTaxonomy
    RiskTaxonomyCreateRequest:
      properties:
        name:
          type: string
          title: Name
        version:
          type: string
          title: Version
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        is_default:
          type: boolean
          title: Is Default
          default: false
      type: object
      required:
      - name
      - version
      title: RiskTaxonomyCreateRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    LocalToken:
      type: apiKey
      in: header
      name: X-Local-Token